PP-5782 Change test to use card number in test BIN range file only
A test that was checking what cardid returns for a card number in
the test BIN ranges file was actually looking for a number that’s
also in Worldpay’s BIN ranges file with slightly different data.
It’s arbitrary which BIN ranges file is used in this case, which
made the test fail sometimes. Make the test use a card number
that’s only in the test BIN ranges file to fix this.